requirements.txt
rembg[cpu]==2.0.65
onnxruntime
numpy
Pillow
